+++ src/pages/Wecome/index.tsx
@@ -1,5 +1,5 @@
 import PageWarp from '@/components/BusinessComponent/PageWarp'
-import React from 'react'
+import React, { useEffect } from 'react'
 
 const Component: React.FC = (): JSX.Element => {
   return (